Step down as a maintainer
As I previously announced, I currently don't have the time to improve this project. As it is an important project for the community, I think I shouldn't give the impression I do maintain the repo or that my opinions should be taken into account if anyone wants to step in and take over the repo. I want to clearly communicate that this is now up to grabs for anyone with at least a little bit of courage and vision.